Update converter.py
Fix issue #81 and #105 properly, by amending the way we handle already-converted choices field from graphene-django 3.0.

The fix introduced in #111 made the error people experienced from the bad handling go away, but it didn't actually set the `is_required` flag correctly. This is because mounted BlankValue Field actually wrapped the underlying choices enum in `NonNull` when pertinent.

Ths issue was related to graphene-django 3.0 taking the converted enum class and mounting it outright in a BlankValueField, instead of just saving the enum itself.

We solve this by, for graphene-django 3.0, by retrieving the wrapped choices Enum, and reusing this.